permission to edit/delete course

+++ src/components/ShowCourse.js
@@ -5,6 +5,7 @@ import Jumbotron from 'react-bootstrap/Jumbotron';
 import Button from 'react-bootstrap/Button';
 import { withRouter } from 'react-router-dom';
 import Table from 'react-bootstrap/Table';
+import Alert from 'react-bootstrap/Alert';
 
 function ShowCourse(props) {
   const [data, setData] = useState({});
@@ -12,13 +13,16 @@ function ShowCourse(props) {
   const apiUrl = "http://localhost:3000/api/courses/" + props.match.params.id;
   const apiStudentListUrl = "http://localhost:3000/api/studentListByCourse/" + props.match.params.id;
   const [studentData, setStudentData] = useState([]);
+  const [sessionId, setSessionId] = useState('');
+  const [message, setMessage] = useState('');
 
   useEffect(() => {
     setShowLoading(false);
     const fetchData = async () => {
       const result = await axios(apiUrl);
       console.log('results from courses', result.data);
-      setData(result.data);
+      setData(result.data.course);
+      setSessionId(result.data.sessionId);
 
       const resultStudentList = await axios(apiStudentListUrl);
       console.log('results from studentList', resultStudentList.data);
@@ -30,27 +34,43 @@ function ShowCourse(props) {
     fetchData();
   }, []);
 
-  const editCourse = (id) => {
-    props.history.push({
-      pathname: '/editCourse/' + id
-    });
+  const editCourse = (id, creatorId) => {
+    console.log("creatorId >>> ", creatorId);
+
+    if (creatorId != sessionId) {
+      console.log("diff session !!");
+      setMessage('You are not allow to edit course !');
+
+    } else {
+      props.history.push({
+        pathname: '/editCourse/' + id
+      });
+    }
 
   };
 
-  const deleteCourse = (id) => {
-    setShowLoading(true);
-    const course = {
-      courseCode: data.courseCode,
-      courseName: data.courseName,
-      section: data.section,
-      semester: data.semester,
-      studentNumber: data.studentNumber
-    };
-    axios.delete(apiUrl, course)
-      .then((result) => {
-        setShowLoading(false);
-        props.history.push('/listCourses')
-      }).catch((error) => setShowLoading(false));
+  const deleteCourse = (id, creatorId) => {
+    console.log("creatorId >>> ", creatorId);
+
+    if (creatorId != sessionId) {
+      console.log("diff session !!");
+      setMessage('You are not allow to delete course !');
+
+    } else {
+      setShowLoading(true);
+      const course = {
+        courseCode: data.courseCode,
+        courseName: data.courseName,
+        section: data.section,
+        semester: data.semester,
+        studentNumber: data.studentNumber
+      };
+      axios.delete(apiUrl, course)
+        .then((result) => {
+          setShowLoading(false);
+          props.history.push('/listCourses')
+        }).catch((error) => setShowLoading(false));
+    }
   };
 
   return (
@@ -58,6 +78,13 @@ function ShowCourse(props) {
       {showLoading && <Spinner animation="border" role="status">
         <span className="sr-only">Loading...</span>
       </Spinner>}
+
+      {message != '' &&
+        <Alert variant="danger">
+          Message : {message}
+        </Alert>
+      }
+
       <Jumbotron>
         <h1>Course Name: {data.courseName}</h1>
         <p>Course Code: {data.courseCode}</p>
@@ -65,8 +92,8 @@ function ShowCourse(props) {
         <p>Semester: {data.semester}</p>
 
         <p>
-          <Button type="button" variant="primary" onClick={() => { editCourse(data._id) }}>Edit</Button>&nbsp;
-          <Button type="button" variant="danger" onClick={() => { deleteCourse(data._id) }}>Delete</Button>
+          <Button type="button" variant="primary" onClick={() => { editCourse(data._id, data.creator) }}>Edit</Button>&nbsp;
+          <Button type="button" variant="danger" onClick={() => { deleteCourse(data._id, data.creator) }}>Delete</Button>
         </p>
       </Jumbotron>
